refactor(controllers\exercise): trocar dominio exercise para challenge
@janinelps @eluvya @Cibely87

132
