+++ src/models/entity/Candidate.ts
@@ -11,7 +11,7 @@ import {
 } from 'typeorm'
 import { IsDate } from 'class-validator'
 import { HiringProcess } from './HiringProcess'
-import { Exercise } from './Exercise';
+import { Challenge } from './Challenge';
 
 @Entity()
 export class Candidate {
@@ -23,12 +23,12 @@ export class Candidate {
   @ManyToOne(() => HiringProcess, hiringProcess => hiringProcess.candidates, { onDelete: 'CASCADE' })
   hiringProcess: HiringProcess
 
-  @OneToOne(() => Exercise, exercise => exercise.candidate, {
+  @OneToOne(() => Challenge, exercise => exercise.candidate, {
     cascade: true,
     onDelete: 'CASCADE'
   })
   @JoinColumn({ name: 'exercise_id' })
-  exercise: Exercise;
+  exercise: Challenge;
 
 
   @Column({ name: 'email', type: 'varchar' })
